+++ src/platform/plugins/shared/dashboard/server/dashboard_saved_object/index.ts
@@ -16,4 +16,6 @@ export type {
   GridData,
   SavedDashboardPanel,
   SavedDashboardSection,
+  StoredControlState,
+  StoredControlGroupInput,
 } from './schema';
