+++ src/platform/plugins/shared/dashboard/server/dashboard_saved_object/schema/v3/index.ts
@@ -12,5 +12,7 @@ export type {
   GridData,
   SavedDashboardPanel,
   SavedDashboardSection,
+  StoredControlState,
+  StoredControlGroupInput,
 } from './types';
 export { dashboardAttributesSchema } from './v3';
