+++ src/platform/plugins/shared/dashboard/server/dashboard_saved_object/schema/v3/types.ts
@@ -8,6 +8,7 @@
  */
 
 import type { TypeOf } from '@kbn/config-schema';
+import type { controlsGroupSchema } from '@kbn/controls-schemas';
 import type { SavedDashboardPanel as SavedDashboardPanelV2 } from '../v2';
 import type { dashboardAttributesSchema, gridDataSchema, sectionSchema } from './v3';
 
@@ -23,3 +24,32 @@ export type SavedDashboardPanel = Omit<SavedDashboardPanelV2, 'gridData'> & { gr
  * A saved dashboard section parsed directly from the Dashboard Attributes
  */
 export type SavedDashboardSection = TypeOf<typeof sectionSchema>;
+
+/**
+ * A saved sticky dashboard control parsed directly from the Dashboard Attributes control panels JSON
+ */
+export type StoredControlState = Pick<
+  TypeOf<typeof controlsGroupSchema>['controls'][number],
+  'grow' | 'type' | 'width'
+> & {
+  order: number; // order is generated from the array order
+  id: string; // id is required
+  dataViewRefName?: string;
+  explicitInput: Omit<
+    TypeOf<typeof controlsGroupSchema>['controls'][number],
+    'grow' | 'type' | 'width'
+  >;
+};
+
+export type StoredControlGroupInput = Omit<
+  TypeOf<typeof dashboardAttributesSchema>['controlGroupInput'],
+  'panelsJSON' | 'ignoreParentSettingsJSON'
+> & {
+  panels: { [key: string]: Omit<StoredControlState, 'id'> };
+  ignoreParentSettings: {
+    ignoreFilters?: boolean;
+    ignoreQuery?: boolean;
+    ignoreTimerange?: boolean;
+    ignoreValidations?: boolean;
+  };
+};
